Changed location of function to utils
