+++ eduiddashboard/saml2/views.py
@@ -1,8 +1,5 @@
 import pprint
 
-from cgi import escape
-from urllib import unquote, quote
-
 from saml2 import BINDING_HTTP_REDIRECT, BINDING_HTTP_POST
 from saml2.client import Saml2Client
 from saml2.metadata import entity_descriptor
@@ -19,7 +16,7 @@ from pyramid.renderers import render_to_response, render
 from pyramid.security import authenticated_userid
 from pyramid.view import view_config, forbidden_view_config
 
-from eduiddashboard.saml2.utils import get_saml2_config, get_location
+from eduiddashboard.saml2.utils import get_saml2_config, get_location, sanitize_url
 from eduiddashboard.saml2.auth import authenticate, login, logout
 from eduiddashboard.saml2.cache import (IdentityCache, OutstandingQueriesCache,
                                         StateCache, )
@@ -119,12 +116,7 @@ def login_view(request):
     login_redirect_url = request.registry.settings.get(
         'saml2.login_redirect_url', '/')
 
-    try:
-        came_from = sanitize_url(request.GET.get('next', login_redirect_url))
-    except UnicodeDecodeError:
-        log.warn('A malicious user tried to crash the application '
-                 'by sending non-unicode input in the next parameter')
-        return HTTPBadRequest("Non-unicode input, please try again.")
+    came_from = sanitize_url(request.GET.get('next', login_redirect_url))
 
     if authenticated_userid(request):
         return HTTPFound(location=came_from)
@@ -366,13 +358,3 @@ def get_authn_request(request, came_from, selected_idp,
     oq_cache = OutstandingQueriesCache(request.session)
     oq_cache.set(session_id, came_from)
     return info
-
-def sanitize_url(url):
-    # If the url is not quoted we only escape it. Otherwise we
-    # have to unquote, escape and quote again before returning it.
-    # The Saml2Client expects the url to be UTF-8 encoded in accordance
-    # with RFC 3986 and therefore we return it as such.
-    if url == unquote(url):
-        return escape(url, quote = True).encode("UTF-8")
-    else:
-        return quote(escape(unquote(url), quote = True)).encode("UTF-8")
